backend/requirements.txt
Django
djangorestframework
djoser
drf_yasg
django-cors-headers
pandas
openpyxl